+++ src/pocketmine/network/protocol/PlayerArmorEquipmentPacket.php
@@ -36,7 +36,7 @@ class PlayerArmorEquipmentPacket extends DataPacket{
 	}
 
 	public function decode(){
-		$this->eid = $this->getInt();
+		$this->eid = $this->getLong();
 		$this->slots[0] = $this->getByte();
 		$this->slots[1] = $this->getByte();
 		$this->slots[2] = $this->getByte();
@@ -45,7 +45,7 @@ class PlayerArmorEquipmentPacket extends DataPacket{
 
 	public function encode(){
 		$this->reset();
-		$this->putInt($this->eid);
+		$this->putLong($this->eid);
 		$this->putByte($this->slots[0]);
 		$this->putByte($this->slots[1]);
 		$this->putByte($this->slots[2]);
