+++ src/pocketmine/network/protocol/TakeItemEntityPacket.php
@@ -42,7 +42,7 @@ class TakeItemEntityPacket extends DataPacket{
 	public function encode(){
 		$this->reset();
 		$this->putInt($this->target);
-		$this->putInt($this->eid);
+		$this->putLong($this->eid);
 	}
 
 }
